+++ components/form/Button.js
@@ -6,7 +6,14 @@ const Button = ({ className, children, color, onClick, outline, Icon }) => {
   const outlineClass = outline ? 'outline' : 'solid';
 
   return (
-    <button className={classNames(styles[outlineClass][color], className)} onClick={onClick}>
+    <button
+      className={classNames(
+        'transition-colors uppercase duration-150 rounded-md focus:shadow-outline inline-flex items-center justify-center ',
+        styles[outlineClass][color],
+        className
+      )}
+      onClick={onClick}
+    >
       {children}
       {Icon && <span className="ml-2 -mr-1">{Icon}</span>}
     </button>
@@ -29,16 +36,14 @@ Button.defaultProps = {
 
 const styles = {
   solid: {
-    primary:
-      'transition-colors uppercase duration-150 bg-amcovad-primary-500 rounded-md focus:shadow-outline hover:bg-amcovad-secondary-600 hover:text-white inline-flex items-center justify-center ',
-    secondary:
-      'text-white transition-colors uppercase duration-150 bg-amcovad-secondary-600 rounded-md focus:shadow-outline hover:bg-amcovad-primary-500 hover:text-white inline-flex items-center justify-center'
+    primary: 'bg-amcovad-primary-500 hover:bg-amcovad-secondary-600 hover:text-white  ',
+    secondary: 'text-white bg-amcovad-secondary-600 hover:bg-amcovad-primary-500 hover:text-white '
   },
   outline: {
     primary:
-      'text-amcovad-secondary-600 uppercase transition-colors duration-150 border border-amcovad-secondary-600 rounded-md focus:shadow-outline hover:bg-amcovad-primary-500 hover:border-amcovad-primary-500 hover:text-white inline-flex items-center justify-center',
+      'text-amcovad-secondary-600 border border-amcovad-secondary-600 hover:bg-amcovad-primary-500 hover:border-amcovad-primary-500 hover:text-white ',
     secondary:
-      'text-amcovad-primary-500 uppercase transition-colors duration-150 border border-amcovad-primary-500 rounded-md focus:shadow-outline hover:bg-amcovad-secondary-600 hover:border-amcovad-primary-500 hover:text-white inline-flex items-center justify-center'
+      'text-amcovad-primary-500 border border-amcovad-primary-500 hover:bg-amcovad-secondary-600 hover:border-amcovad-primary-500 hover:text-white '
   }
 };
