+++ data/dashboard/transaction/steps.js
@@ -1,0 +1,17 @@
+import StepOne from '@/components/dashboard/Transaction/StepOne';
+import StepTwo from '@/components/dashboard/Transaction/StepTwo';
+
+const data = [
+  {
+    id: 1,
+    label: 'Quick Start',
+    content: <StepOne />
+  },
+  {
+    id: 2,
+    label: 'Set Terms & Conditions',
+    content: <StepTwo />
+  }
+];
+
+export default data;
